bootloader: POSIX: install signal handlers before forking the child process
In POSIX onefile codepath, have the parent process install signal
handlers (to forward or ignore received signals) *before* forking
the child process instead of after it.

This aims to mitigate sporadic failures of our signal handling test
(`test_onefile_signal_handling`); there, we signal the parent process
from the child, and depending on the scenario (forward/block) expect
child to either receive the forwarded signal or not receive the signal.
If the signal handlers are installed after the `fork()` and `execvp()`,
it might happen (in CPU contention scenarios) that the child proceeds
with execution while the parent stalls for a bit, and thus child
ends up signalling the parent before the latter has set up its signal
handlers.

In practice, this early setup should not make any difference; while
the signal handlers are inherited when the child process is forked,
the subsequent program re-execution via execvp() ends up resetting
them.

In addition, have the signal-forwarding handler check that child
PID is valid (!= 0), and have the stored PID reset to 0 after
the child exits. This ensures that the signal handler is no-op
for the interval after the handlers are installed and before the
child is forked, and for the interval after the child exits and
before the signal handlers are uninstalled. This implicitly assumes
that storage of PID value is atomic and no further synchronization
is needed (which should be true on contemporary platforms).