bootloader: add a retry loop for deleting onefile temp. dir. on Windows
Add a retry loop for deleting temporary directory of onefile
applications on Windows.

This aims to mitigate situations when a spawned grandchild process
is still in the process of being shut down. If their executable is
part of the frozen application, it and its dependencies will still
be locked at the time when we attempt to clean them up.

This seems to happen with `QtWebEngineProcess.exe` helper process
that is spawned by QtWebEngine-based applications; the onefile
QtWebEngine tests on our CI seem to be sporadically failing to
remove the helper executable and the DLLs its loads in the case
of increased CPU contention (e.g., increased number of pytest
workers).

Similarly, it seems that when running an x64 application under an
arm64 system, the OS ends up locking (some of?) the python extensions
and DLLs for a short while, and interferes with the cleanup.