loader: clean-up/simplify PyiFrozenLoader implementation
As the `PyiFrozenLoader` instance is now tied to specific module,
we can further simplify implementation in its methods.

Specifically, we can pass the PYZ entry name and is_package flag
via PyiFrozenLoader's constructor, and avoid re-computing the name
and looking up the entry in each and every method. Similarly, we
can determine the module file path in constructor, store it into
`path` attribute, and have the `get_filename` and other methods
read it from there.