conftest: turn compiled_dylib fixture into session-scoped fixture
We need to compile the test dynamic library only once per test
session, instead of in each test that makes use of it. However,
in order to make the `compiled_dylib` fixture session-scoped, we
need to move it from `PyInstaller/utils/conftest.py` into
`tests/functional/conftest.py`.

This is because its data directory needs to be resolved based on
the module's location (as ˙request.path` is not available in
session-scoped fixtures). And even if we were to use  test-scoped
fixture and infer location from `request.path`, the fixture would
be valid only for test files from `tests/functional` directory, so
it makes more sense to put it into its local `conftest.py` file.