tests: remove pyi_get_datadir module
Remove the `pyi_get_datadir` module, since none of the test
programs are using it anymore.

Instead of using `pyi_get_datadir.get_data_dir()` in the test
program code, the tests should either copy the data files into
frozen application (which is what would normally happen), or pass
the path to the temporary data directory to the test program via
command-line arguments (e.g., if data is to be processed). Both
cases will make it immediately obvious that the test is in fact
making use of the `data_dir` fixture (with test program using
`pyi_get_datadir.get_data_dir()` the test did not seem to be
using the fixture in spite of including it).