bootloader: fix calloc() result check to account for AIX behavior
On AIX, when `calloc` is called with number of elements set to 0,
it returns a NULL pointer (unless `_LINUX_SOURCE_COMPAT` is defined,
but this is negated by having `_ALL_SOURCE` defined, which we do).
This differs from behavior on other platforms, where a non-NULL
address is returned in such cases, and NULL is returned only on
allocation failure.

Therefore, perform a NULL check on returned pointer to detect
allocation failure only when we call `calloc` with non-zero
number of elements.

This fixes spurious run-time error on AIX when no Wflags and/or no
Xflags are specified via bootloader's run-time options.